+++ loans-service/src/main/java/ru/avalc/loansservice/model/Customer.java
@@ -20,7 +20,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "customer_id")
-    private int customerId;
+    private long customerId;
 
     private String name;
