+++ loans-service/src/main/java/ru/avalc/loansservice/repo/LoansRepository.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Repository
 public interface LoansRepository extends CrudRepository<Loan, Long> {
 
-    List<Loan> findByCustomerIdOrderByStartDateDesc(int customerId);
+    List<Loan> findByCustomerIdOrderByStartDateDesc(long customerId);
 }
